pyproject.toml
[project]
name = "homework"
description = "Recruitment task"
authors = [{name="Jakub Płachecki", email="jakub.plachecki@gmail.com"}]
readme = "README.md"
version = "0.0.1"

[tool.poetry]
package-mode = false
version = "0.0.1"

[tool.poetry.dependencies]
python = "^3.11"
alembic = "1.15.1"
alembic-postgresql-enum = "1.7.0"
asyncpg= "0.29.0"
fastapi = {version = "0.115.2", extras = ["all"]}
SQLAlchemy = "2.0.30"
uvicorn = "0.29.0"

[tool.poetry.group.dev.dependencies]
asgi-lifespan = "2.1.0"
coverage = "^7.1.0"
gevent = "^23.9.1"
isort = "=5.13.2" # match .pre-commit-config.yaml
pip-audit = "^2.7.3"
polyfactory = "2.16.0"
pre-commit = "^3.7.1"
pytest = "^7.4"
pytest-asyncio = "^0.21"
pytest-cov = "^4.1"
pytest-env = "^1.1.3"
pytest-mock = "^3.14.0"
python-multipart = "^0.0.18"
python-dotenv = "^1.0.1"
ruff = "=0.9.6" # match .pre-commit-config.yaml

[tool.coverage.run]
concurrency = ["gevent"]
branch = true
source = ["app", "tests"]
omit = ["app/schemas/factories.py", "tests/*"]

[tool.coverage.report]
fail_under = 50
show_missing = true
skip_covered = true
omit = ["app"]

[tool.mypy]
strict = true
files = ["app", "tests"]

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
log_cli_format = "%(asctime)s %(levelname)s %(message)s"

env = [
    "DB_HOST=db:5432",
    "DB_INSTANCE_NAME=postgres",
    "DB_USER=postgres",
    "DB_PASSWD=postgres",
]
markers = [
    "online: Marks test as online",
]

[tool.isort]
profile = "black"
use_parentheses = true
line_length = 99
multi_line_output = 3
skip = ["venv/", ".venv/"]
extend_skip = [".json", ".yml", ".yaml", ".toml", ".md", ".txt", ".rst", ".ini", ".cfg", ".conf", ".env"]
include_trailing_comma = true
force_grid_wrap = 2

[tool.ruff]
lint.select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I001",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "S",  # flake8-bandit
]
lint.ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "F405",  # usage of things from * imports
    "S101",  # asserts usages
    "B904",
    "I001",
]



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
